Example PR: auto provision a S3 static website
This is an example PR (do not merge!). It is a template to show
how to get the CI system to automatically provision a static website
with an S3 bucket and cloudfront.

Instance creation workflow:

* Scientist work with IT to determine if this setup is appropriate
  for their use case.
* (Optional) Register a domain.
* Request a SSL certificate from the AWS certificate manager
* Scientist create a PR similar to this one. Create a new file
with a unique stack_name and updated parameters.
* Have someone review and approve the PR.
* Merge the PR
* The CI system will auto provision the resources
* An email will be sent to the owner email address with website endpoint.
* The newly created resources will contain tags similear to these:

```
key: Department
value: Platform

key: OwnerEmail
value: joe.smith@sagebase.org

key: Project
value: Infrastructure

```

To remove the resources:

* Instance owner must create a new PR that deletes the files that was added by
the PR that was used to add the instance.
* Have an an AWS account Admin review your change.
* Approve and merge.