[IT-937] Update deny policy on htan bucket (#372)
The deny policy on htan bucket policy for s3-synapse-sync buckets
causes access denied on changes to owner.txt file so we need
to remove it.

Our CFN template creates a bucket with an owner.txt file then applies
the bucket policy which includes a deny unless the PUT includes
the S3 canonical identifiers.  The file is added on initial bucket
creation.  However when the s3objects-macro attempts to update owner.txt
file it gets a access denied because the macro doesn't do a PUT with
the identifiers.