[IT-937] Fix bucket policy
The strategy for this fix attempt is to deny S3UserARNs
who do not provide canonical account IDs. This works
because S3UserARNs are the only users that have access
to the bucket anyways.

The Previous attempt to fix owner.txt file update with
BucketServiceRoleId condition failed so we remove it..