[IT-937] add required AWS lambda permission (#375)
The previous build failed because the NotificationConfiguration requires
a `AWS::Lambda::Permission` which has not been setup yet.  When
attempting to add that permission and NotificationConfiguration at the
same time we run into a circular dependency problem which is documented
in the NotificationConfiguration docs[1].

The doc is basically saying that the only way to avoid the circular
dependency is to deploy in two steps.  The imnplication is that every
time we create a bucket we would need to deploy the AWS::Lambda::Permission
first then run another deploy to setup the NotificationConfiguration.

This step is o setup the permission.

[1] https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-s3-bucket-notificationconfig.html