[IT-937] Create s3 synapse sync bucket (#379)
Change yaml to jinja template to allow us to deploy bucket
notification on a parameter condition

The synapse sync bucket template is specific to the synapse sync
lambda so it makes sense to move it to https://github.com/Sage-Bionetworks/s3-synapse-sync
repo.  Now we can reference the template from there and
use it in this repo.

This will create a synapse s3 sync bucket without enabling bucket
notifications. To enable bucket notification set
EnableNotificationConfiguration to true.

depends on Sage-Bionetworks/scicomp-provisioner#378