Provision miniDREAM 2021 infrastructure (#478)
I'm deploying infrastructure for miniDREAM 2021. This stack is composed of a high-memory EC2 instance in a private subnet behind by a public load balancer (both using existing aws-infra templates).

We will be hosting RStudio on port 8787 of this instance, and we will setup an nginx reverse proxy to redirect incoming HTTPS (port 443) traffic from the ALB to port 8787. The ALB will take care of upgrading HTTP (port 80) traffic to HTTPS (port 443).

We have a custom setup to integrate with Synapse evaluation queues, so we can't use the RStudio packer AMI.