Update Bruno's admin role with local viewer (#530)
This PR is partially motivated by IT-1967 and partially motivated by the need to inspect the bucket policy, which can only be done by IAM identities within the same account as the S3 bucket. As a result, my current sandbox role doesn't have access to the bucket policy. Since I have access to the viewer role in scicomp, I was thinking of granting access to my specific assumed-role at the resource/bucket level. If this works, then we can migrate others to using the same viewer role to administer/access the HTAN buckets.