[IT-1967] Cleanup bucket access (#540)
We are removing access to sandbox and scicomp AWS accounts
which means that the ARN for those users will no longer be
valid.  We are switching to use another AWS account (identity-central)
to allow cross account access to buckets.  This change is
to update user access to scicomp buckets.

* Remove users that are no longer Sage employees
* Update ARN to reference users in identity-central account.

Now users must first login to org-sagebase-identitycentral AWS
account to access scicomp buckets.