Create nf-syn23664726-s3-temp (#548)
This bucket is a temporary bucket to try and troubleshoot issues that Nebula Genomics is having with transferring a subset of files to `nf-syn23664726-s3`, we plan to try and transfer those files to this new bucket ( nf-syn23664726-s3-temp) and see if we can consolidate everything into nf-syn23664726-s3 on our end. After that this bucket can be deleted.