grant bucket access (#560)
I added my ARN to the `grantAccess` list in order to get access to this bucket. I believe it contains files related to data usage stats that we need for the AD Portal. Larsson is listed as the bucket owner but has not accessed it himself; he says it's fine for me to be added.