Create ec2 and alb for miniDREAM 2023 (#571)
Setting up a stack to run miniDREAM this year for about 16 students. Similar to [last year](https://github.com/Sage-Bionetworks/scicomp-provisioner/pull/522), the stack is composed of a high-memory EC2 instance in a private subnet behind by a public load balancer (both using existing aws-infra templates), which will be used to provide a RStudio environment for students taking the miniDREAM courses.

This PR was built using #522 as a reference.