[IT-3128]  Remove service user
This repo is old and has been archived.  The service user that was setup
for it was used by travis-ci.  This is no longer valid so we are
removing the service user.  If we need to revive this repo then we
should move it to github action and use OIDC.