remove region restriction from PEC and CMC buckets (#658)
The Psychencode DACC needs to download roughly 2 TB of data from the CMC and PEC project, but they do not have access to AWS. This PR removes the region restriction to allow for the DACC to complete a key final step of the migration from Synapse. Egress cost should be minimal for this download (~$180), and only a few select number of users still have access to these projects on Synapse.